GHC 8 compatibility, and compatibility with newer build-depends. Fixes #3
The `instance Lift Text` is already implicitly imported,
because `haskell-src-meta` imports from `th-orphans`, which
imports from `th-lift-instances` which has this instance.

In order to resolve the duplicate instance, this commit
adds `th-lift-instances` directly to the `build-depends`,
making the dependency explicit.

The `global` TH function is deprecated, and according to the
corresponding deprecation warning, `varE` should be used;
according to what I've found, it exists in all versions of
template-haskell.